+++ lib/Config.php
@@ -8,7 +8,7 @@ use ProcessWire\InputfieldWrapper,
 /**
  * SearchEngine Config
  *
- * @version 0.3.0
+ * @version 0.4.0
  * @author Teppo Koivula <teppo.koivula@gmail.com>
  * @license Mozilla Public License v2.0 http://mozilla.org/MPL/2.0/
  */
@@ -195,9 +195,42 @@ class Config extends Base {
             $compatible_fieldtypes->value = $data[$compatible_fieldtypes->name] ?? $options[$compatible_fieldtypes->name] ?? null;
             $compatible_fieldtypes->notes = $this->_('Please note that selecting fieldtypes not selected by default may result in various problems. Change these values only if you\'re sure that you know what you\'re doing.');
         }
+        $compatible_fieldtypes->notes .= $this->getCompatibleFieldtypeDiff($compatible_fieldtypes->value);
         $advanced_settings->add($compatible_fieldtypes);
 
         return $fields;
+    }
+
+    /**
+     * Get a list of changes (additions and removals) made to the compatible fieldtypes setting
+     *
+     * @param array $compatible_fieldtypes Current list of compatible fieldtypes.
+     * @return string String representation of the changes.
+     */
+    protected function getCompatibleFieldtypeDiff(array $compatible_fieldtypes): string {
+
+        // get a diff by comparing module default setting value and current setting value
+        $base = \ProcessWire\SearchEngine::$defaultOptions['compatible_fieldtypes'];
+        $diff = array_filter([
+            'added' => implode(', ', array_diff($compatible_fieldtypes, $base)),
+            'removed' => implode(', ', array_filter(array_diff($base, $compatible_fieldtypes), function($fieldtype) {
+                return $this->wire('modules')->isInstalled($fieldtype);
+            })),
+        ]);
+
+        // construct output string
+        $out = "";
+        if (!empty($diff)) {
+            $out .= "\n";
+            if (!empty($diff['added'])) {
+                $out .= "\n+ " . sprintf($this->_('Added fieldtypes: %s'), $diff['added']);
+            }
+            if (!empty($diff['removed'])) {
+                $out .= "\n- " . sprintf($this->_('Removed fieldtypes: %s'), $diff['removed']);
+            }
+        }
+
+        return $out;
     }
     
 }